==> 00_baseline.ipynb <==
import pandas as pd
from pathlib import Path

data_dir = Path('data')

sb = pd.read_csv(data_dir/'submission_format.csv', index_col='id')

sb.head()

sb.loc[:,0:5] = pd.read_csv(data_dir/'train_labels.csv').groupby(['verified']).mean()[-1:].values

sb.head()

len(sb)

sb.to_csv(data_dir/'submissions/mean_baseline.csv')
