==> 10_silver_QC_20191127.ipynb <==
## Predict St Lucia Metal Roofs

Using rotated to hz + OpenCv border


%reload_ext autoreload
%autoreload 2
%matplotlib inline

from fastai.vision import *
import geopandas as gpd
from resizeimage import resizeimage
import datetime
import uuid
from os import listdir
from os.path import isfile, join
from ipyexperiments import *

data_dir = Path('data')
img_size = '256'

TRAIN_IMAGES=data_dir/f'train/rotated/clipped/{img_size}'
SILVER_TRAIN=TRAIN_IMAGES/'silver_train/all_unverified'
#pseudo test set - what we run inference on
SILVER_UNVERIFIED = TRAIN_IMAGES/'silver_unverified'

train_images=data_dir/f'train/rotated/clipped/{img_size}'
test_images=data_dir/f'test/rotated/clipped/{img_size}'

silver_names = get_image_files(SILVER_TRAIN)

len(silver_names)

#see nb 04_5
df_silver=pd.read_csv(data_dir/'df_all_repl_st_lucia_castries_gold_concrete_cement.csv')

df_silver.drop(columns=['target'],inplace=True)

len(df_silver)

df_silver.head()

df_all=pd.read_csv(data_dir/'df_train_all.csv')

len(df_all)

silver_ids=df_silver.id.values.tolist()

train_ids=set(df_all.id.values.tolist())

unmatched = train_ids.symmetric_difference(set(silver_ids)) 

unmatched

unmatched_df=df_all.loc[df_all['id'].isin(unmatched)]

unmatched_df.head()

len(unmatched_df)

df_unverified=df_all.loc[df_all['verified']==False]

len(df_unverified)

df_verified=df_all.loc[df_all['verified']==True]

len(df_verified)
